requirement.txt
--find-links=https://download.pytorch.org/whl/torch_stable.html
pip-review
wheel 
pytest
tqdm
numpy
torch==1.9.1+cu111
torchvision==0.10.1+cu111
torchaudio==0.9.1 
opencv-python
jupyter
notebook
requests
plotly

attrdict
pyyaml
colorlog
black
black[jupyter]
